sqlite: creates "makeDispatcherNative"
