+++ src/adapters/sqlite/type.js
@@ -4,13 +4,24 @@ import { type ConnectionTag } from '../../utils/common'
 import { type ResultCallback } from '../../utils/fp/Result'
 
 import type { RecordId } from '../../Model'
-import type { TableName, SchemaVersion } from '../../Schema'
+import type { AppSchema, TableName, SchemaVersion } from '../../Schema'
+import type { SchemaMigrations } from '../../Schema/migrations'
 
 import { type DirtyFindResult, type DirtyQueryResult } from '../common'
 
 export type SQL = string
 export type SQLiteArg = string | boolean | number | null
 export type SQLiteQuery = [SQL, SQLiteArg[]]
+
+export type SQLiteAdapterOptions = $Exact<{
+  dbName?: string,
+  schema: AppSchema,
+  migrations?: SchemaMigrations,
+  synchronous?: boolean,
+  experimentalUseJSI?: boolean,
+}>
+
+export type DispatcherType = 'asynchronous' | 'synchronous' | 'jsi'
 
 export type NativeBridgeBatchOperation =
   | ['execute', TableName<any>, SQL, SQLiteArg[]]
